Fixes Non-canonical cast (double) is deprecated in PHP 8.5

+++ src/Handshake/RequestVerifier.php
@@ -45,7 +45,7 @@ class RequestVerifier {
      * @return bool
      */
     public function verifyHTTPVersion($val): bool {
-        return 1.1 <= (double)$val;
+        return 1.1 <= (float)$val;
     }
 
     /**
